crawler/requirements.txt
# Core dependencies
# Use newer FastAPI version that works with newer Pydantic
fastapi>=0.100.0,<0.101.0
# Update uvicorn to a version that works with newer httptools
uvicorn[standard]>=0.20.0,<0.21.0
# Remove httpx as it's included with uvicorn[standard]
# httpx>=0.19.0,<0.20.0
backoff>=1.11.1,<2.0.0
prometheus-client>=0.11.0,<0.12.0
# Update aioredis to a newer version that fixes TimeoutError conflict
redis>=4.5.0,<5.0.0  # Use redis instead of aioredis as it now includes async support
# Update qdrant-client to a newer version
qdrant-client

# ML and NLP dependencies
sentence-transformers>=2.2.2,<3.0.0
# Use newer Pydantic version
pydantic>=1.9.0,<2.0.0
numpy>=1.24.3,<2.0.0
# Update spacy to a newer version
spacy>=3.5.0,<4.0.0
# Use matching spacy model version
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
transformers>=4.35.0,<5.0.0

# Database and caching
networkx>=3.2.1,<4.0.0

# Utilities
python-dotenv>=0.19.0,<1.0.0
prometheus-fastapi-instrumentator>=5.7.1,<6.0.0
aiohttp>=3.8.5,<4.0.0
PyMuPDF>=1.21.0,<2.0.0
pyyaml>=6.0.1,<7.0.0
psutil>=5.9.5,<6.0.0
beautifulsoup4>=4.9.3,<5.0.0
gitpython>=3.1.0,<4.0.0

# Remove explicit httptools version as it will be handled by uvicorn[standard]
# httptools==0.5.0

# Build dependencies
setuptools>=65.5.1
wheel>=0.38.4

# Ollama/Mistral integration
langchain
langchain-community
ollama>=0.1.6
PyPDF2>=3.0.0

# Legal compliance and dictionary processing
scancode-toolkit>=32.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl 